+++ exo_k/ktable5d.py
@@ -314,7 +314,7 @@ class Ktable5d(Data_table):
         self.Ng=self.weights.size
 
         conversion_factor=u.Unit(grid_p_unit).to(u.Unit('Pa'))
-        self.logpgrid=np.array(logpgrid)+np.log10(conversion_factor)
+        self.logpgrid=np.array(logpgrid, dtype=float)+np.log10(conversion_factor)
         self.pgrid=10**self.logpgrid #in Pa
         self.p_unit='Pa'
         self.Np=self.logpgrid.size
